% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/between.R
\name{between}
\alias{\%btwn\%}
\alias{between}
\title{Between}
\usage{
x \%btwn\% rng

between(x, rng, inclusive = TRUE)
}
\arguments{
\item{x}{Numeric vector (or single value): the values to check if they're
between \code{rng}.}

\item{rng}{Two-element numeric cector: the range.}

\item{inclusive}{Whether or not to include the lower and upper values of
the range (default TRUE).}
}
\value{
A logical vector of the same length as \code{x}.

  For every element in \code{x}, return \code{TRUE} if the value lies within
  \code{rng}, and \code{FALSE} otherwise.
}
\description{
Determine if a numeric value is between the specified range
}
\details{
\code{x \%btwn\% rng} is equivalent to \code{between(x, rng, inclusive = TRUE)}.
}
\examples{
5 \%btwn\% c(1, 10)
c(5, 20) \%btwn\% c(5, 10)
between(5, c(5, 10))
between(5, c(5, 10), inclusive = FALSE)
}

