% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfFactorize.R
\name{dfFactorize}
\alias{dfFactorize}
\title{Factorize data.frame}
\usage{
dfFactorize(df, only = c(), ignore = c())
}
\arguments{
\item{df}{A data.frame.}

\item{only}{A vector of column names. Only convert these columns to factors.}

\item{ignore}{A vector of column names. Do not convert these columns.}
}
\value{
A data.frame with all character columns (or only a subset of them
  if the parameters were set) converted to factors.
}
\description{
Convert character columns in a data.frame to factors.
}
\details{
Given a data.frame, convert character columns to factors.  By default,
all character columns are converted, but the user can specify which columns
to convert or to not convert.
}
\note{
If any of the names in \code{only} or \code{ignore} are not valid
  columns, an error is raised.

Only one of \code{only} or \code{ignore} can be used in a single call.
}
\examples{
df <- data.frame(a = 1:3, b = letters[1:3], c = LETTERS[1:3],
                 stringsAsFactors = FALSE)
str(df)
str(dfFactorize(df))
str(dfFactorize(df, only = "b"))
str(dfFactorize(df, ignore = "b"))
str(dfFactorize(df, ignore = c("b", "c")))
str(dfFactorize(df, only = c("a")))

# The following examples result in errors
\dontrun{str(dfFactorize(df, only = c("z")))}
\dontrun{str(dfFactorize(df, only = "b", ignore = "c"))}
}

