% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggExtra.R
\name{ggExtra}
\alias{ggExtra}
\alias{ggMarginal}
\alias{plotCount}
\alias{removeGrid}
\alias{removeGridX}
\alias{removeGridY}
\alias{rotateTextX}
\title{Useful ggplot2 functions and layers - deprecated}
\usage{
removeGrid(...)

removeGridX(...)

removeGridY(...)

rotateTextX(...)

plotCount(...)

ggMarginal(...)
}
\arguments{
\item{...}{ignored}
}
\description{
These functions have graduated from \code{rsalad} and are now available in
the \code{\link[ggExtra]{ggExtra}} package. You can download \code{ggExtra} from
\href{http://cran.r-project.org/web/packages/ggExtra/index.html}{CRAN} or
\href{https://github.com/daattali/ggExtra}{GitHub}.
}

