% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggplotLayers}
\alias{ggplotLayers}
\alias{removeGrid}
\alias{removeGridX}
\alias{removeGridY}
\alias{rotateTextX}
\title{Useful ggplot2 layers}
\usage{
removeGrid(x = TRUE, y = TRUE)

removeGridX()

removeGridY()

rotateTextX()
}
\arguments{
\item{x}{Whether to apply th layer to the X axis.}

\item{y}{Whether to apply th layer to the Y axis.}
}
\value{
A ggplot2 layer that can be added to an existing ggplot2 object.
}
\description{
A small set of useful layers to add to ggplot2 plots.
}
\details{
\code{removeGrid} always removes the minor gridlines from the plot, and
removes the major gridlines from X and/or Y axis (both by default).

\code{removeGridX} is a shortcut for \code{removeGrid(x = TRUE, y = FALSE)}

\code{removeGridY} is a shortcut for \code{removeGrid(x = FALSE, y = TRUE)}

\code{rotateTextX} causes the labels on the x axis to be rotated 90 degrees
so that they are vertical.  This can be useful when there are many labels
along the x axis and having them horizontal causes them to overlap.
}
\note{
The \code{ggplot2} package is required for these functions.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  df <- data.frame(x = 1:50, y = 1:50)
  p <- ggplot2::ggplot(df, ggplot2::aes(x, y)) + ggplot2::geom_point()
  p + removeGrid()
  p + removeGrid(y = FALSE)
  p + removeGridX()

  df <- data.frame(x = paste("letter", LETTERS, sep = "_"),
                   y = seq(length(LETTERS)))
  p <- ggplot2::ggplot(df, ggplot2::aes(x, y)) + ggplot2::geom_point()
  p + rotateTextX()
}
}

