% Generated by roxygen2 (4.0.2): do not edit by hand
\name{move}
\alias{move}
\alias{moveBack}
\alias{moveBack_}
\alias{moveFront}
\alias{moveFront_}
\title{Move columns to front/back}
\usage{
moveFront(df, ...)

moveBack(df, ...)

moveFront_(df, cols)

moveBack_(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{...}{Comma separated list of unquoted expressions. Used in NSE
functions.}

\item{cols}{A vector of column names to move. Used in SE functions.}
}
\value{
A data.frame with the given columns moved either to the beginning
  or end of the data.frame.
}
\description{
Move specific columns in data.frames to be the first or last columns.
}
\details{
\code{moveFront} moves columns to be the first columns in a data.frame,
\code{moveBack} moves columns to be the last columns in a data.frame.

The order of the columns as passed in the arguments will be the
same order that the columns will appear in the resulting data.frame,
regardless if the front or back functions are called.

\code{moveFront} and \code{moveBack} are the non-standard evaluation (NSE)
versions, while \code{moveFront_} and \code{moveBack_} are the standard
evaluation (SE) versions.  Look at the examples to see the difference in
usage.
}
\note{
The \code{dplyr} package is required for these functions.

If any of the column names given are not valid columns, an error
  is raised.
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, c = 7:9, d = 10:12)
moveFront(df, d)
moveFront(df, d, c)
moveFront(df, c, d)
moveBack(df, b)
moveBack_(df, "a")
moveFront_(df, c("b", "d"))
}

