% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/notIn.R
\name{notIn}
\alias{\%nin\%}
\alias{notIn}
\title{Not in}
\usage{
x \%nin\% y

notIn(x, y)
}
\arguments{
\item{x}{Vector or NULL: the values to be matched.}

\item{y}{Vector or NULL: the values to be matched against.}
}
\value{
A logical vector of the same length as \code{x}.

  For every element in \code{x}, return \code{FALSE} if the value exists
  in \code{y}, and \code{TRUE} otherwise.
}
\description{
Determine if values in the first argument don't exist in the second argument.
}
\details{
\code{\%nin\%} is the functional inverse of \code{\%in\%}.

\code{lhs \%nin\% rhs} is equivalent to \code{notIn(lhs, rhs)}.
}
\examples{
"a" \%nin\% letters
"a" \%nin\% LETTERS
c("a", "A") \%nin\% letters
notIn("A", letters)
}

