% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCount.R
\name{plotCount}
\alias{plotCount}
\title{Plot count data}
\usage{
plotCount(x)
}
\arguments{
\item{x}{A data.frame. See 'Details' for more information.}
}
\value{
A ggplot2 layer that can be added to an existing ggplot2 object.
}
\description{
Create a bar plot of count data in a data.frame.
}
\details{
The argument to this function is expected to be a data.frame with two
columns: the first column is a list of unique values, the second column
is a list of integers representing count data for the first column.

This function is meant to be called with the result of \code{\link{dfCount}},
though any data.frame with a proper structure will also work.
}
\note{
The \code{ggplot2} package is required for these functions.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  if (requireNamespace("nycflights13", quietly = TRUE)) {
    flights <- nycflights13::flights
    df <- dfCount(flights, "origin")
    plotCount(df)
  }

  plotCount(dfCount(infert, "education", sort = FALSE))
  plotCount(table(infert$education))
}
}
\seealso{
\code{\link{dfCount}}
}

