% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rdeanutil}
\alias{rdeanutil}
\alias{rdeanutil-package}
\title{rdeanutil}
\description{
Collection of useful utility functions for R
}
\details{
This package provides a collections of useful R functions that
I've found myself reusing often in many different projects.  There isn't
really a common specific theme to all these functions, but they are all
related to either manipulating data frames or helping with productivity.
}

