% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsalad.R
\docType{package}
\name{rsalad}
\alias{rsalad}
\alias{rsalad-package}
\title{rsalad}
\description{
A mix of useful R functions that are good for you
}
\details{
Collection of functions that help with easily manipulating data.frames, such
as converting variables to factors, rearranging variables,and converting list
of lists to data.frames.
}

