% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiffsym.R
\name{setdiffsym}
\alias{setdiffsym}
\title{Symmetric set difference}
\usage{
setdiffsym(x, y)
}
\arguments{
\item{x, y}{Vectors}
}
\value{
All elements that are in one vector but not the other
}
\description{
Perform symmetric set difference, as opposed to \code{base::setdiff} which
performs asymmetric set difference.
}
\examples{
setdiffsym(1:4, 3:5)
setdiffsym(3:5, 1:4)
}

