% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setupSpinMyRtest.R
\name{setupSpinMyRtest}
\alias{setupSpinMyRtest}
\title{Set up a test directory to experiment with \code{spinMyR}}
\usage{
setupSpinMyRtest(loc = getwd())
}
\arguments{
\item{loc}{The path where you want to set up the test directories.}
}
\value{
The path to the root of the test directory.
}
\description{
This function creates a few directories that try to mimic a real
data-analysis project structure, and adds a simple R script and a data file.
After setting up these files and directories, you can run both
\code{knitr::spin} and \code{rsalad::spinMyR} on the R script to see the
benefits of \code{spinMyR}.
}
\details{
The console output from this function will give more specific insturctions on
how to interact with this test directory.
}
\examples{
\dontrun{
library(rsalad)
tmp <- setupSpinMyRtest("~")
origwd <- setwd(tmp)
knitr::spin("container/R/spinMyRtest.R")
setupSpinMyRtest("..")
spinMyR("R/spinMyRtest.R", wd = "container")
setupSpinMyRtest("..")
spinMyR("R/spinMyRtest.R", wd = "container",
        outDir = "output", figDir = "coolplots")
setwd(origwd)
unlink(tmp, recursive = TRUE, force = TRUE)
}
}
\seealso{
\code{\link[rsalad]{spinMyR}}

\code{\link[knitr]{spin}}
}

