% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinMyR.R
\name{spinMyR}
\alias{spinMyR}
\title{Create markdown/HTML reports from R script with no hassle - deprecated}
\usage{
spinMyR(...)
}
\arguments{
\item{...}{ignored}
}
\description{
This functions has graduated from \code{rsalad} and is now available in
the \code{\link[ezknitr]{ezknitr}} package. You can download \code{ezknitr} from
\href{http://cran.r-project.org/web/packages/ezknitr/index.html}{CRAN} or
\href{https://github.com/daattali/ezknitr}{GitHub}.
}

