% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rsalad.R
\docType{package}
\name{rsalad}
\alias{rsalad}
\alias{rsalad-package}
\title{rsalad}
\description{
A mix of useful R functions that are good for you
}
\details{
Collection of functions that help with easily manipulating data.frames (e.g.
converting variables to factors, rearranging variables), improving ggplot2
plots (e.g. removing certain gridlines, rotating axis labels), and improve
usability and functionality of knitr::spin (e.g. better handling of working
directories and output directories).
}

