% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rsalad}
\alias{rsalad}
\alias{rsalad-package}
\title{rsalad}
\description{
A mix of commonly useful R functions, mostly for data.frames and ggplot2
}
\details{
This package provides a collection of useful R functions that I've found
myself reusing often in many different projects.  While there isn't a
common specific theme to all these functions, they are mostly either for
manipulating data frames or improving ggplot2 plots.
}

